Another attempt to fix timing calculation on arduino playback

+++ main.c
@@ -153,7 +153,7 @@ void loop(){\n\
             delayMicroseconds(%u);\n\
           }\n\
         }\n\
-}\n", (unsigned int) ((1000000 / 44100) * output_bitrate_divisor));
+}\n", (unsigned int) ( (float)1000000 / 44100 * output_bitrate_divisor ) );
 
 
 		}
